backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pypdf==5.1.0
python-docx==1.1.2
